Finished get single exercise by ID, tests written and passed, postman tested fine.
