+++ .gitignore
@@ -66,3 +66,4 @@ yarn.lock
 
 # ignore sqlite3 database
 *.db3
+index.js
