+++ exercises/exercisesModels.js
@@ -1,0 +1,12 @@
+const db = require('../data/dbConfig');
+
+function findExerciseById(id) {
+  return db('exercises')
+    .select()
+    .where({ id })
+    .first();
+}
+
+module.exports = {
+  findExerciseById,
+};
