--- routes/exercises/exercise-helpers.js
@@ -1,12 +1,0 @@
-const db = require('../../models/dbConfig');
-
-function findExerciseById(id) {
-  return db('exercises')
-    .select()
-    .where({ id: id })
-    .first();
-}
-
-module.exports = {
-  findExerciseById,
-};
